Move skin_changer files to its own folder
