+++ src/NullHooks.vcxproj
@@ -124,7 +124,7 @@
     <ClCompile Include="core\features\visuals\no_flash.cpp" />
     <ClCompile Include="core\features\misc\spectator_list.cpp" />
     <ClCompile Include="core\features\visuals\player_esp.cpp" />
-    <ClCompile Include="core\features\visuals\skin_changer.cpp" />
+    <ClCompile Include="core\features\visuals\skin_changer\skin_changer.cpp" />
     <ClCompile Include="core\hooks\functions\alloc_key_values_memory.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">NotUsing</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">NotUsing</PrecompiledHeader>
@@ -248,7 +248,7 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="core\features\misc\backtrack.hpp" />
-    <ClInclude Include="core\features\visuals\skin_changer.hpp" />
+    <ClInclude Include="core\features\visuals\skin_changer\skin_changer.hpp" />
     <ClInclude Include="core\hooks\functions\event_globals.hpp" />
     <ClInclude Include="core\features\visuals\models.hpp" />
     <ClInclude Include="dependencies\interfaces\i_physics_surface_props.hpp" />
